pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "shark-cli"
version = "0.1.0"
description = "A professional CLI for the Sharktech Cloud Provider API"
authors = ["Sharktech Community"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "shark_cli" }]
keywords = ["sharktech", "cloud", "cli", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
click = "^8.1"
httpx = "^0.27"
pyyaml = "^6.0"
rich = "^13.0"

[tool.poetry.scripts]
shark = "shark_cli.main:cli"
